Extract collectedTestsData to separate function
